+++ cli-tool/test/commands/export.rb
@@ -39,14 +39,20 @@ class TestCommandsExport < MTest::Unit::TestCase
     export_command = Taylor::Commands::Export.new(['--dry-run'], {})
     assert_include export_command.puts_data, 'docker run'
     assert_include export_command.puts_data, File.join(Dir.pwd, 'exports')
-    assert_include export_command.puts_data, "hellrok/taylor:v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
   end
 
   def test_check_docker_command_override_export_directory
     export_command = Taylor::Commands::Export.new(['--dry-run'], { export_directory: '/tmp/exports' })
     assert_include export_command.puts_data, 'docker run'
     assert_include export_command.puts_data, '/tmp/exports'
-    assert_include export_command.puts_data, "hellrok/taylor:v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
   end
 
   def test_check_docker_command_set_build_cache
@@ -56,6 +62,9 @@ class TestCommandsExport < MTest::Unit::TestCase
     assert_include export_command.puts_data, 'docker run'
     assert_include export_command.puts_data, File.join(Dir.pwd, 'exports')
     assert_include export_command.puts_data, '/tmp/build'
-    assert_include export_command.puts_data, "hellrok/taylor:v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:windows-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:osx-v#{TAYLOR_VERSION}"
+    assert_include export_command.puts_data, "hellrok/taylor:web-v#{TAYLOR_VERSION}"
   end
 end